+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - Add support for Tautulli as a way to connect to Plex
+- New config option `exclude_tag` to ignore series by Sonarr tag.
+
 
 ## [1.2.0] - 2025-09-11